requirements.txt
aiogram>=3.0.0
httpx>=0.24.0
pydantic>=2.0.0
structlog>=23.0.0
apscheduler>=3.10.0
google-generativeai>=0.3.0
openai>=1.0.0
langdetect>=1.0.9
feedparser>=6.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
ruff>=0.1.0
mypy>=1.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
prometheus-client>=0.17.0
sentry-sdk>=1.30.0
locust>=2.15.0 